% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.validann}
\alias{plot.validann}
\title{Plot ANN validation results.}
\usage{
\method{plot}{validann}(results, obs, sim, type = c("multi", "single"), ...)
}
\arguments{
\item{results}{object of class `validann' as returned
by \code{\link{validann}}. This is a list comprising metrics and
statistics that can be used for validating ANN models.}

\item{obs,sim}{vectors comprising observed (\code{obs}) and simulated
(\code{sim}) examples of a single response variable used for computing
\code{results} object.}

\item{type}{character string defining how plots should be displayed.
The default is ``multi'' where multiple plots are displayed together
according to whether they are used for predictive, replicative or
structural validation. For ``single'', each plot is displayed on its own.
If the session is interactive, the user will be asked to confirm a new
page whether \code{type} is ``single'' or ``multi''}
}
\description{
Plot method for objects of class `validann'. Produces a series
   of plots used for validating and assessing ANN models based on results
   returned by \code{\link{validann}}.
}
\details{
This function can be invoked by calling
   \code{plot(results, obs, sim)} for an object \code{results} of class
   `validann'.

   To produce plots for all (predictive, replicative and structural)
   types of validation metrics and statistics, \code{results$predictive},
   \code{results$replicative} and \code{results$structural} must be
   \code{TRUE} and, for replicative and structural validation, corresponding
   results must have been successfully computed by \code{\link{validann}}
   and returned in object \code{results}.

   If \code{results$predictive} is \code{TRUE}, a scatter plot, Q-Q plot and
   time/sample plot of observed (\code{obs}) versus predicted (\code{sim})
   data are produced.

   If \code{results$replicative} is \code{TRUE} and \code{results$residuals}
   is not \code{NULL}, plots of the model residuals are produced including
   histogram, Q-Q plot (standardized residuals compared to standard normal),
   autocorrelation (acf), partial autocorrelation (pacf), standardized
   residual versus predicted output (i.e. \code{sim}) and standardized
   residual versus time/order of the data.

   If \code{results$structural} is \code{TRUE} and results$y_hat is not
   \code{NULL}, plots of the model residuals
   are produced including histogram, Q-Q plot (standardized residuals
   compared to standard normal), autocorrelation (acf),
   partial autocorrelation (pacf), standardized residuals versus predicted
   output (i.e. \code{sim}) and standardized residual versus time/order of
   the data.

   Setting \code{results$predictive}, \code{results$replicative} and/or
   \code{results$structural} to \code{FALSE} prior to calling
   \code{plot.validann} will turn off the respective validation plots.
}
\examples{
## Build ANN model and compute replicative and structural validation results
data("ar9")
samp <- sample(1:1000, 200)
y <- ar9[samp, ncol(ar9)]
x <- ar9[samp, -ncol(ar9)]
x <- x[, c(1,4,9)]
fit <- ann(x, y, size = 1, act_hid = "tanh", act_out = "linear", rang = 0.1)
results <- validann(fit, x = x)
obs <- observed(fit)
sim <- fitted(fit)

## Plot replicative and structural validation results to the current device
## - a single page for each type of validation
plot(results, obs, sim)

## Plot results to the current device - a single page for each plot
plot(results, obs, sim, type = "single")

## Plot replicative and structural validation results to single file
pdf("RepStructValidationPlots.pdf")
plot(results, obs, sim)
dev.off()

## Get predictive validation results for above model based on a new sample
## of ar9 data.
samp <- sample(1:1000, 200)
y <- ar9[samp, ncol(ar9)]
x <- ar9[samp, -ncol(ar9)]
x <- x[, c(1,4,9)]
obs <- y
sim <- predict(fit, newdata = x)$values
results <- validann(fit, obs = obs, sim = sim, x = x)

## Plot predictive results only to file
results$replicative <- results$structural <- FALSE
pdf("PredValidationPlots.pdf")
plot(results, obs, sim)
dev.off()
}
\seealso{
\code{\link{validann}}
}

