% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ann.R
\name{predict.ann}
\alias{predict.ann}
\title{Predict new examples using a trained neural network.}
\usage{
\method{predict}{ann}(object, newdata)
}
\arguments{
\item{object}{an object of class `ann' as returned by function \code{ann}.}

\item{newdata}{matrix, data frame or vector of input data. A vector is
considered to comprise examples of a single input or predictor variable.}
}
\value{
a list with components:
\item{values}{matrix of values returned by the trained ANN.}
\item{derivs}{matrix of derivatives of hidden (columns \code{1:object$size})
and output (final column) nodes.}
}
\description{
Predict new examples using a trained neural network.
}
\details{
This function is a method for the generic function \code{predict()}
for class `ann'. It can be invoked by calling \code{predict(x)} for an
object \code{x} of class `ann', or directly by calling
\code{predict.ann(x)} regardless of the class of \code{x}.

Derivatives returned are useful for sensitivity analyses.
}
\examples{
## Select new sample from ar9 dataset and use ann object `fit' to predict y.
## ---
samp <- sample(1:1000, 200)
y <- ar9[samp, ncol(ar9)]
x <- ar9[samp, -ncol(ar9)]
x <- x[, c(1,4,9)]
sim <- predict(fit, newdata = x)$values
obs <- y
}
\seealso{
\code{\link{ann}}
}

